responder: 📵 move `dispense()` logic into a submodule
